+++ Core/DataInputUtility.cs
@@ -71,19 +71,19 @@ namespace Greatbone.Core
             return null;
         }
 
-        public static D StringToObject<D>(string v, int proj = 0) where D : IData, new()
+        public static D StringToObject<D>(string v, short proj = 0) where D : IData, new()
         {
-            JObj jo = (JObj)new JsonParse(v).Parse();
+            JObj jo = (JObj) new JsonParse(v).Parse();
             return jo.ToObject<D>(proj);
         }
 
-        public static D[] StringToArray<D>(string v, int proj = 0) where D : IData, new()
+        public static D[] StringToArray<D>(string v, short proj = 0) where D : IData, new()
         {
-            JArr ja = (JArr)new JsonParse(v).Parse();
+            JArr ja = (JArr) new JsonParse(v).Parse();
             return ja.ToArray<D>(proj);
         }
 
-        public static string ToString<D>(D v, int proj = 0) where D : IData
+        public static string ToString<D>(D v, short proj = 0) where D : IData
         {
             JsonContent cont = new JsonContent(false, true, 4 * 1024);
             cont.Put(null, v, proj);
@@ -92,7 +92,7 @@ namespace Greatbone.Core
             return str;
         }
 
-        public static string ToString<D>(D[] v, int proj = 0) where D : IData
+        public static string ToString<D>(D[] v, short proj = 0) where D : IData
         {
             JsonContent cont = new JsonContent(false, true, 4 * 1024);
             cont.Put(null, v, proj);
@@ -101,7 +101,7 @@ namespace Greatbone.Core
             return str;
         }
 
-        public static string ToString<D>(List<D> v, int proj = 0) where D : IData
+        public static string ToString<D>(List<D> v, short proj = 0) where D : IData
         {
             JsonContent cont = new JsonContent(false, true, 4 * 1024);
             cont.Put(null, v, proj);
@@ -137,12 +137,12 @@ namespace Greatbone.Core
             return null;
         }
 
-        public static D FileToObject<D>(string file, int proj = 0) where D : IData, new()
+        public static D FileToObject<D>(string file, short proj = 0) where D : IData, new()
         {
             try
             {
                 byte[] bytes = File.ReadAllBytes(file);
-                JObj jo = (JObj)new JsonParse(bytes, bytes.Length).Parse();
+                JObj jo = (JObj) new JsonParse(bytes, bytes.Length).Parse();
                 if (jo != null)
                 {
                     return jo.ToObject<D>(proj);
@@ -155,12 +155,12 @@ namespace Greatbone.Core
             return default(D);
         }
 
-        public static D[] FileToArray<D>(string file, int proj = 0) where D : IData, new()
+        public static D[] FileToArray<D>(string file, short proj = 0) where D : IData, new()
         {
             try
             {
                 byte[] bytes = File.ReadAllBytes(file);
-                JArr ja = (JArr)new JsonParse(bytes, bytes.Length).Parse();
+                JArr ja = (JArr) new JsonParse(bytes, bytes.Length).Parse();
                 if (ja != null)
                 {
                     return ja.ToArray<D>(proj);
@@ -173,12 +173,12 @@ namespace Greatbone.Core
             return null;
         }
 
-        public static List<D> FileToList<D>(string file, int proj = 0) where D : IData, new()
+        public static List<D> FileToList<D>(string file, short proj = 0) where D : IData, new()
         {
             try
             {
                 byte[] bytes = File.ReadAllBytes(file);
-                JArr ja = (JArr)new JsonParse(bytes, bytes.Length).Parse();
+                JArr ja = (JArr) new JsonParse(bytes, bytes.Length).Parse();
                 if (ja != null)
                 {
                     return ja.ToList<D>(proj);
