+++ Core/Form.cs
@@ -136,7 +136,7 @@ namespace Greatbone.Core
             return false;
         }
 
-        public bool Get<D>(string name, ref D v, int proj = 0) where D : IData, new()
+        public bool Get<D>(string name, ref D v, short proj = 0) where D : IData, new()
         {
             throw new NotImplementedException();
         }
@@ -190,17 +190,17 @@ namespace Greatbone.Core
             throw new NotImplementedException();
         }
 
-        public bool Get<D>(string name, ref D[] v, int proj = 0) where D : IData, new()
+        public bool Get<D>(string name, ref D[] v, short proj = 0) where D : IData, new()
         {
             throw new NotImplementedException();
         }
 
-        public bool Get<D>(string name, ref List<D> v, int proj = 0) where D : IData, new()
+        public bool Get<D>(string name, ref List<D> v, short proj = 0) where D : IData, new()
         {
             throw new NotImplementedException();
         }
 
-        public D ToObject<D>(int proj = 0) where D : IData, new()
+        public D ToObject<D>(short proj = 0) where D : IData, new()
         {
             D dat = new D();
             dat.ReadData(this, proj);
@@ -226,12 +226,12 @@ namespace Greatbone.Core
             throw new NotImplementedException();
         }
 
-        public D[] ToArray<D>(int proj = 0) where D : IData, new()
+        public D[] ToArray<D>(short proj = 0) where D : IData, new()
         {
             throw new NotImplementedException();
         }
 
-        public List<D> ToList<D>(int proj = 0) where D : IData, new()
+        public List<D> ToList<D>(short proj = 0) where D : IData, new()
         {
             throw new NotImplementedException();
         }
