+++ Core/FormContent.cs
@@ -360,7 +360,7 @@ namespace Greatbone.Core
             throw new NotImplementedException();
         }
 
-        public FormContent Put(string name, IData v, int proj = 0, string Label = null, string Help = null, bool ReadOnly = false, bool required = false)
+        public FormContent Put(string name, IData v, short proj = 0, string Label = null, string Help = null, bool ReadOnly = false, bool required = false)
         {
             if (name != null)
             {
@@ -384,7 +384,7 @@ namespace Greatbone.Core
             return this;
         }
 
-        public FormContent Put<D>(string name, D[] v, int proj = 0, string Label = null, string Help = null, bool ReadOnly = false, bool required = false) where D : IData
+        public FormContent Put<D>(string name, D[] v, short proj = 0, string Label = null, string Help = null, bool ReadOnly = false, bool required = false) where D : IData
         {
             if (name != null)
             {
@@ -410,7 +410,7 @@ namespace Greatbone.Core
             return this;
         }
 
-        public FormContent Put<D>(string name, List<D> v, int proj = 0, string Label = null, string Help = null, bool ReadOnly = false, bool required = false) where D : IData
+        public FormContent Put<D>(string name, List<D> v, short proj = 0, string Label = null, string Help = null, bool ReadOnly = false, bool required = false) where D : IData
         {
             if (name != null)
             {
