+++ Core/IDataInput.cs
@@ -36,17 +36,17 @@ namespace Greatbone.Core
 
         bool Get(string name, ref Dictionary<string, string> v);
 
-        bool Get<D>(string name, ref D v, int proj = 0) where D : IData, new();
+        bool Get<D>(string name, ref D v, short proj = 0) where D : IData, new();
 
-        bool Get<D>(string name, ref D[] v, int proj = 0) where D : IData, new();
+        bool Get<D>(string name, ref D[] v, short proj = 0) where D : IData, new();
 
-        bool Get<D>(string name, ref List<D> v, int proj = 0) where D : IData, new();
+        bool Get<D>(string name, ref List<D> v, short proj = 0) where D : IData, new();
 
-        D ToObject<D>(int proj = 0) where D : IData, new();
+        D ToObject<D>(short proj = 0) where D : IData, new();
 
-        D[] ToArray<D>(int proj = 0) where D : IData, new();
+        D[] ToArray<D>(short proj = 0) where D : IData, new();
 
-        List<D> ToList<D>(int proj = 0) where D : IData, new();
+        List<D> ToList<D>(short proj = 0) where D : IData, new();
 
         ///
         /// Write a single (or current) data entry into the given output object.
